[Refactor] Split Profile module views and models + fix exit alert layout
